requirements.txt
pandas
numpy
matplotlib
shapely
xmltodict
requests
python-dotenv